+++ docking/ui/geometry.py
@@ -675,10 +675,10 @@ def _build_item_geometries(
             hide_cross=hide_cross,
         )
         draw_rect = Rect(
-            int(math.floor(draw_x)),
-            int(math.floor(draw_y)),
-            max(1, int(math.ceil(scaled_size))),
-            max(1, int(math.ceil(scaled_size))),
+            math.floor(draw_x),
+            math.floor(draw_y),
+            max(1, math.ceil(scaled_size)),
+            max(1, math.ceil(scaled_size)),
         )
         anchor_x, anchor_y = _item_anchor(
             pos=pos, draw_x=draw_x, draw_y=draw_y, scaled_size=scaled_size
@@ -802,13 +802,13 @@ def _compute_item_hover_base_rect(
     background_rect: Rect,
     theme: Theme,
 ) -> Rect:
-    item_padding = max(0, int(round(theme.item_padding)))
-    top_padding = max(0, int(math.ceil(theme.top_padding)))
-    bottom_padding = max(0, int(math.ceil(theme.bottom_padding)))
+    item_padding = max(0, round(theme.item_padding))
+    top_padding = max(0, math.ceil(theme.top_padding))
+    bottom_padding = max(0, math.ceil(theme.bottom_padding))
 
     if is_horizontal(pos=pos):
-        left = int(math.floor(draw_rect.x - item_padding / 2))
-        right = int(math.ceil(draw_rect.x + draw_rect.w + item_padding / 2))
+        left = math.floor(draw_rect.x - item_padding / 2)
+        right = math.ceil(draw_rect.x + draw_rect.w + item_padding / 2)
         if pos == Position.BOTTOM:
             top = draw_rect.y - top_padding
             bottom = draw_rect.y + draw_rect.h + bottom_padding
@@ -816,8 +816,8 @@ def _compute_item_hover_base_rect(
             top = draw_rect.y - bottom_padding
             bottom = draw_rect.y + draw_rect.h + top_padding
     else:
-        top = int(math.floor(draw_rect.y - item_padding / 2))
-        bottom = int(math.ceil(draw_rect.y + draw_rect.h + item_padding / 2))
+        top = math.floor(draw_rect.y - item_padding / 2)
+        bottom = math.ceil(draw_rect.y + draw_rect.h + item_padding / 2)
         if pos == Position.LEFT:
             left = draw_rect.x - bottom_padding
             right = draw_rect.x + draw_rect.w + top_padding
@@ -857,11 +857,11 @@ def _compute_item_hit_rect(
     background_rect: Rect,
     theme: Theme,
 ) -> Rect:
-    main_pad = max(0, int(round(theme.item_padding / 2)))
+    main_pad = max(0, round(theme.item_padding / 2))
     if is_horizontal(pos=pos):
         center_x = draw_rect.x + draw_rect.w / 2
         width = draw_rect.w + main_pad
-        left = int(round(center_x - width / 2))
+        left = round(center_x - width / 2)
         right = left + max(1, width)
         top = min(draw_rect.y, background_rect.y)
         bottom = max(draw_rect.y + draw_rect.h, background_rect.y + background_rect.h)
@@ -869,7 +869,7 @@ def _compute_item_hit_rect(
 
     center_y = draw_rect.y + draw_rect.h / 2
     height = draw_rect.h + main_pad
-    top = int(round(center_y - height / 2))
+    top = round(center_y - height / 2)
     bottom = top + max(1, height)
     left = min(draw_rect.x, background_rect.x)
     right = max(draw_rect.x + draw_rect.w, background_rect.x + background_rect.w)
@@ -887,7 +887,7 @@ def _compute_main_axis_boundaries(
         starts = [min(static_dock_rect.x, background_rect.x)]
         centers = [hover_rect.x + hover_rect.w / 2 for hover_rect in hover_rects]
         for left_center, right_center in pairwise(centers):
-            starts.append(int(round((left_center + right_center) / 2)))
+            starts.append(round((left_center + right_center) / 2))
         starts.append(
             max(
                 static_dock_rect.x + static_dock_rect.w,
@@ -899,7 +899,7 @@ def _compute_main_axis_boundaries(
     starts = [min(static_dock_rect.y, background_rect.y)]
     centers = [hover_rect.y + hover_rect.h / 2 for hover_rect in hover_rects]
     for top_center, bottom_center in pairwise(centers):
-        starts.append(int(round((top_center + bottom_center) / 2)))
+        starts.append(round((top_center + bottom_center) / 2))
     starts.append(
         max(
             static_dock_rect.y + static_dock_rect.h,
@@ -919,7 +919,7 @@ def _compute_background_rect(
     drop_gap: float,
 ) -> Rect:
     gap = max(0, int(theme.distance_from_edge))
-    shelf_cross = max(1, int(round(theme.shelf_height)))
+    shelf_cross = max(1, round(theme.shelf_height))
     stroke_width = max(0.0, float(theme.stroke_width))
     main_padding = theme.item_padding + 2 * theme.h_padding + 4 * stroke_width
 
@@ -938,8 +938,8 @@ def _compute_background_rect(
             (draw_rects[-1].w if is_horizontal(pos=pos) else draw_rects[-1].h)
             + main_padding
         ) / 2
-        main_start = int(round(first_center - first_half))
-        main_end = int(round(last_center + last_half + drop_gap))
+        main_start = round(first_center - first_half)
+        main_end = round(last_center + last_half + drop_gap)
     else:
         if is_horizontal(pos=pos):
             main_start = static_dock_rect.x
@@ -957,27 +957,21 @@ def _compute_background_rect(
 
     if pos == Position.BOTTOM:
         shelf_slide = hide_offset * (shelf_cross + gap)
-        shelf_y = int(
-            round(
-                static_dock_rect.y
-                + static_dock_rect.h
-                - gap
-                - shelf_cross
-                + shelf_slide
-            )
+        shelf_y = round(
+            static_dock_rect.y + static_dock_rect.h - gap - shelf_cross + shelf_slide
         )
         return Rect(main_start, shelf_y, max(1, main_end - main_start), shelf_cross)
     if pos == Position.TOP:
         shelf_slide = hide_offset * (shelf_cross + gap)
-        shelf_y = int(round(static_dock_rect.y + gap - shelf_slide))
+        shelf_y = round(static_dock_rect.y + gap - shelf_slide)
         return Rect(main_start, shelf_y, max(1, main_end - main_start), shelf_cross)
     if pos == Position.LEFT:
         shelf_slide = hide_offset * (shelf_cross + gap)
-        shelf_x = int(round(static_dock_rect.x + gap - shelf_slide))
+        shelf_x = round(static_dock_rect.x + gap - shelf_slide)
         return Rect(shelf_x, main_start, shelf_cross, max(1, main_end - main_start))
     shelf_slide = hide_offset * (shelf_cross + gap)
-    shelf_x = int(
-        round(static_dock_rect.x + static_dock_rect.w - gap - shelf_cross + shelf_slide)
+    shelf_x = round(
+        static_dock_rect.x + static_dock_rect.w - gap - shelf_cross + shelf_slide
     )
     return Rect(shelf_x, main_start, shelf_cross, max(1, main_end - main_start))
